requirements.txt
yfinance
mplfinance
pandas
numpy
matplotlib
quantstats
riskfolio-lib